Fix msfvenom check
Commit #60ea3bc broke the application with an unexpected indent error. After fixing the indent error, the new code caused the application to print the output of 'msfvenom -h' and 'please install metasploit.' before the rest of the program could be allowed to execute.  I slightly tweaked and moved the code to a more appropriate place in the application.